+++ fili-core/src/main/java/com/yahoo/bard/webservice/druid/model/datasource/TableDataSource.java
@@ -2,6 +2,7 @@
 // Licensed under the terms of the Apache license. Please see LICENSE.md file distributed with this work for terms.
 package com.yahoo.bard.webservice.druid.model.datasource;
 
+import com.yahoo.bard.webservice.data.config.names.DataSourceName;
 import com.yahoo.bard.webservice.table.ConstrainedTable;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
@@ -13,6 +14,7 @@ import java.util.Set;
  */
 public class TableDataSource extends DataSource {
 
+    private final DataSourceName name;
     /**
      * Constructor.
      *
@@ -20,10 +22,16 @@ public class TableDataSource extends DataSource {
      */
     public TableDataSource(ConstrainedTable physicalTable) {
         super(DefaultDataSourceType.TABLE, physicalTable);
+        Set<DataSourceName> dataSourceNames = physicalTable.getDataSourceNames();
+        if (dataSourceNames.size() != 1) {
+            throw new IllegalArgumentException("TableDataSource can only be used with single datasource name tables.");
+        }
+
+        this.name = dataSourceNames.stream().findFirst().get();
     }
 
     public String getName() {
-        return getPhysicalTable().getTableName().asName();
+        return name.asName();
     }
 
     @Override