+++ fili-core/src/main/java/com/yahoo/bard/webservice/table/availability/ConcreteAvailability.java
@@ -2,13 +2,11 @@
 // Licensed under the terms of the Apache license. Please see LICENSE.md file distributed with this work for terms.
 package com.yahoo.bard.webservice.table.availability;
 
-import com.yahoo.bard.webservice.data.config.names.TableName;
+import com.yahoo.bard.webservice.data.config.names.DataSourceName;
 import com.yahoo.bard.webservice.metadata.DataSourceMetadataService;
 import com.yahoo.bard.webservice.table.resolver.PhysicalDataSourceConstraint;
 import com.yahoo.bard.webservice.util.SimplifiedIntervalList;
 
-import java.util.Collections;
-import java.util.Map;
 import java.util.Objects;
 import java.util.Set;
 
@@ -17,37 +15,18 @@ import javax.validation.constraints.NotNull;
 /**
  * An availability that provides column and table available interval services for concrete physical table.
  */
-public class ConcreteAvailability implements Availability {
-
-    private final TableName name;
-    private final DataSourceMetadataService metadataService;
-
-    private final Set<TableName> dataSourceNames;
-
+public class ConcreteAvailability extends BaseMetadataAvailability {
     /**
      * Constructor.
      *
-     * @param tableName  The name of the table and data source associated with this Availability
+     * @param dataSourceName  The name of the data source associated with this Availability
      * @param metadataService  A service containing the datasource segment data
      */
     public ConcreteAvailability(
-            @NotNull TableName tableName,
+            @NotNull DataSourceName dataSourceName,
             @NotNull DataSourceMetadataService metadataService
     ) {
-        this.name = tableName;
-        this.metadataService = metadataService;
-
-        this.dataSourceNames = Collections.singleton(name);
-    }
-
-    @Override
-    public Set<TableName> getDataSourceNames() {
-        return dataSourceNames;
-    }
-
-    @Override
-    public Map<String, SimplifiedIntervalList> getAllAvailableIntervals() {
-        return metadataService.getAvailableIntervalsByTable(name);
+        super(dataSourceName, metadataService);
     }
 
     @Override
@@ -63,35 +42,23 @@ public class ConcreteAvailability implements Availability {
                 .map(physicalName -> getAllAvailableIntervals().getOrDefault(
                         physicalName,
                         new SimplifiedIntervalList()
-                )).reduce(SimplifiedIntervalList::intersect).orElse(new SimplifiedIntervalList());
-    }
-
-    /**
-     * Returns the name of the table and data source associated with this Availability.
-     *
-     * @return the name of the table and data source associated with this Availability
-     */
-    protected TableName getName() {
-        return name;
+                ))
+                .reduce(SimplifiedIntervalList::intersect)
+                .orElse(new SimplifiedIntervalList());
     }
 
     @Override
     public String toString() {
-        return String.format(
-                "ConcreteAvailability for table: %s with data source names %s",
-                name.asName(),
-                dataSourceNames
-        );
+        return String.format("ConcreteAvailability for data source: %s", getDataSourceName().asName());
     }
 
     @Override
     public boolean equals(Object obj) {
         if (obj instanceof ConcreteAvailability) {
             ConcreteAvailability that = (ConcreteAvailability) obj;
-            return Objects.equals(name.asName(), that.name.asName())
-                    && Objects.equals(dataSourceNames, that.dataSourceNames)
+            return Objects.equals(getDataSourceName().asName(), that.getDataSourceName().asName())
                     // Since metadata service is mutable, use instance equality to ensure table equality is stable
-                    && metadataService == that.metadataService;
+                    && getDataSourceMetadataService() == that.getDataSourceMetadataService();
         }
         return false;
     }
@@ -99,6 +66,6 @@ public class ConcreteAvailability implements Availability {
     @Override
     public int hashCode() {
         // Leave metadataService out of hash because it is mutable
-        return Objects.hash(name.asName(), dataSourceNames);
+        return Objects.hash(getDataSourceName());
     }
 }